Minor tweaks to backdrop, moves content from modal.js into confirmation.js.
